refactored - breaken by classes added eslint
